// server/src/store/ducks/types/Viagem.ts
export interface Viagem {
    via_os: string;
    via_mot_id: string;
    via_nomePassageiro: string;
    via_telPassageiro: string;
    via_end_origem: string;
    via_end_destino: string;
    via_dataHora_solicitacao: string;
    via_dataHora_embarque: string;
    via_observacao: string;
    via_status: string;
    solicitante: string;
    empresa_viagem: string;
    empresa_solicitante: string;
}